switched to better algrorithm and restructured the classes
