+++ canvas_project/editor/templates/editor/editor.html
@@ -32,6 +32,7 @@
                 "projectSettingsManager": "{% static 'js/projectSettingsManager.mjs' %}",
                 "createCommands": "{% static 'js/createCommands.mjs' %}",
                 "commandPrompt": "{% static 'js/commandPrompt.mjs' %}",
+                "promptCommands": "{% static 'js/promptCommands.mjs' %}",
                 "bootstrap": "{% static 'js/bootstrap.esm.js' %}",
                 "@popperjs/core": "https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/esm/popper.min.js"
             }
