+++ canvas_project/jsconfig.json
@@ -30,6 +30,7 @@
             "projectSettingsManager": ["static/js/projectSettingsManager.mjs"],
             "createCommands": ["static/js/createCommands.mjs"],
             "commandPrompt": ["static/js/commandPrompt.mjs"],
+            "promptCommands": ["static/js/promptCommands.mjs"],
             "bootstrap": ["static/js/bootstrap.esm.js"]
         }
     }
